refactor(structure text form): migrate to typescript
